% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpparmadillotest-package.R
\docType{package}
\name{rcpparmadillotest-package}
\alias{rcpparmadillotest}
\alias{rcpparmadillotest-package}
\title{rcpparmadillotest: Testing cpp11armadillo}
\description{
Testing cpp11armadillo. I run this within a tiny package to test the compiled code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/cpp11armadillo}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp11armadillo/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
